[SPARK-22131] Mesos driver secrets. The driver launches executors that have access to env or file-based secrets.
